Moved logs scheduler to a dedicated package and added a bunch of comments
